requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
python-multipart==0.0.20
groq==0.13.1
pandas==2.2.3
python-dotenv==1.0.1
pydantic==2.10.4
